🔧 chore: Fix metadata.bias increment in Brain class
In the `src/brain.class.js` file, there was an issue with the increment of `metadata.bias` in the `getVertex` method of the `Brain` class. The existing code mistakenly used `bias` instead of `metadata.bias`. This commit fixes the problem by correctly adding `metadata.bias` to the existing value.

The relevant code changes are as follows:
```diff
-    if (metadata.bias) this.definitions[collection][id].metadata.bias += bias
+    if (metadata.bias) this.definitions[collection][id].metadata.bias += metadata.bias
```

Now, when `metadata.bias` is present, it will be properly added to the current value of `metadata.bias` for the specified vertex in the `this.definitions` object.

This commit addresses the issue and ensures correct behavior for the `getVertex` method in the `Brain` class.